+++ CHANGELOG.md
@@ -27,6 +27,15 @@ Additional labels for pre-release metadata:
 * rc.x: stable release candidate.
 
 
+0.2.2 - Unreleased
+------------------
+
+Changed
+
+* User can choose p2p listening port. Previously the port was forced to random.
+* Always generate random ed25519 p2p keypair.
+
+
 0.2.1 - 04-11-2021
 ------------------